yAxis en pourcentage
